Add files via upload
This is an .ods file (compatible with libre calc, and probably excel) that aids in understanding and calculating subnets. Credit for the concepts here goes to The Cyber Mentor, upon whose work this file is generally based. On the bits section of the spreadsheet, if you change the bits from 1 to 0 or back again, then the portions of the subnet mask, hosts, CIDR range, and 2^ calculations all change accordingly.